Recursive Include Support
Right now, the support for `!include` only works at the first level and
only supports including yaml (and json) files.

This change should make it so that if you have a raml file with an
include that pulls in a file that also has an include, it'll keep
walking that tree.

Also, it looks at the file extension and if it isn't `yaml`, `raml` or
`json`, it assumes you just want to pull it in as a string, rather than
make any attempt to parse it.