Change ref_uri to always be an absolute path.
I had left a file called `example.json` in the root of the folder which
was causing false positives. This is because urilib.urlopen() assumes
the relative path is relative to where the test is run rather than a
custom defined directory. Therefore, we change the url to always be an
absolute path as a work-around.